requirements.txt
pyrogram~=2.0.106
tgcrypto
python-dotenv~=1.0.0
rich~=13.4.2